Position添加父子关系
Position添加父子关系

+++ app/src/main/java/com/eflake/keyanimengine/keyframe/IEFAnim.java
@@ -23,9 +23,9 @@ public interface IEFAnim {
 
     boolean isDone();
 
-    int getViewPortPosX(int currentPosX);
+    float getViewPortPosX(float currentPosX);
 
-    int getViewPortPosY(int currentPosY);
+    float getViewPortPosY(float currentPosY);
 
     void setName(String name);
